Refactor msg handlers in Mir struct
- Split mir.go fns into appropriate files
- Removed init and var for msg handlers
- Added a Go routine on msg received
