+++ TASKS.md
@@ -20,7 +20,15 @@
 
 - [ ] MCP Server for Mir
   - [ ] https://www.utcp.io  
-- [ ] TinyGo
+- [ ] DeviceSDK v2
+  - [ ] TinyGo
+  - [x] Refactor
+  - [ ] Store with prefix ID
+  - [ ] Metrics gopsutil
+  - [ ] Bbolt integration (choose between that and badger)
+  - [ ] Tool to see data in badger
+  - [ ] Disk size limits as well as TTL
+  - [ ] Error msg to core eg: disk full
 - [ ] Cockpit
   - [x] New cmd for refresh schema with button next to Delete
   - [x] cfg
